+++ src/stamp-preferences.c
@@ -45,6 +45,8 @@ struct _StampPreferences {
   AdwSwitchRow *play_incoming_sound;
   AdwSpinRow *refresh_interval;
   AdwSwitchRow *important_first;
+  AdwSwitchRow *starred_first;
+  AdwSpinRow *starred_limit;
   AdwPreferencesGroup *accounts_group;
   AdwPreferencesGroup *signatures_group;
   AdwActionRow *add_signature;
@@ -130,6 +132,8 @@ stamp_preferences_class_init (StampPreferencesClass *klass)
   gtk_widget_class_bind_template_child (widget_class, StampPreferences, play_incoming_sound);
   gtk_widget_class_bind_template_child (widget_class, StampPreferences, refresh_interval);
   gtk_widget_class_bind_template_child (widget_class, StampPreferences, important_first);
+  gtk_widget_class_bind_template_child (widget_class, StampPreferences, starred_first);
+  gtk_widget_class_bind_template_child (widget_class, StampPreferences, starred_limit);
   gtk_widget_class_bind_template_child (widget_class, StampPreferences, accounts_group);
   gtk_widget_class_bind_template_child (widget_class, StampPreferences, signatures_group);
   gtk_widget_class_bind_template_child (widget_class, StampPreferences, add_signature);
@@ -310,6 +314,8 @@ stamp_preferences_init (StampPreferences *self)
   g_signal_connect_object (self->background_notifications, "notify::active", G_CALLBACK (on_background_notifications), self, G_CONNECT_DEFAULT);
   g_settings_bind (STAMP_SETTINGS_MAIL, STAMP_PREFS_MAIL_LOAD_BIMI_IMAGES, self->bimi_images, "active", G_SETTINGS_BIND_DEFAULT);
   g_settings_bind (STAMP_SETTINGS_MAIL, STAMP_PREFS_MAIL_IMPORTANT_FIRST, self->important_first, "active", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind (STAMP_SETTINGS_MAIL, STAMP_PREFS_MAIL_STARRED_FIRST, self->starred_first, "active", G_SETTINGS_BIND_DEFAULT);
+  g_settings_bind (STAMP_SETTINGS_MAIL, STAMP_PREFS_MAIL_STARRED_LIMIT, self->starred_limit, "value", G_SETTINGS_BIND_DEFAULT);
   g_settings_bind (STAMP_SETTINGS_MAIL, STAMP_PREFS_MAIL_MARK_READ_TIMEOUT, self->mark_read, "value", G_SETTINGS_BIND_DEFAULT);
   g_settings_bind (STAMP_SETTINGS_MAIL, STAMP_PREFS_MAIL_REFRESH_INTERVAL, self->refresh_interval, "value", G_SETTINGS_BIND_DEFAULT);